requirements.txt
fastapi
pydantic
tortoise-orm
passlib
bcrypt
python-multipart
PyJWT